+++ src/hooks/useImageUpload.ts
@@ -1,0 +1,46 @@
+// src/hooks/useImageUpload.ts
+import { useState } from 'react';
+
+export function useImageUpload() {
+  const [selectedImages, setSelectedImages] = useState<File[]>([]);
+  const [imagePreviewUrls, setImagePreviewUrls] = useState<string[]>([]);
+
+  const handleImageChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const files = Array.from(e.target.files || []);
+    if (files.length + selectedImages.length > 5) {
+      alert('최대 5개의 이미지만 업로드할 수 있습니다.');
+      return;
+    }
+
+    setSelectedImages(prev => [...prev, ...files]);
+
+    // 미리보기 URL 생성
+    files.forEach(file => {
+      const url = URL.createObjectURL(file);
+      setImagePreviewUrls(prev => [...prev, url]);
+    });
+  };
+
+  const removeImage = (index: number) => {
+    setSelectedImages(prev => prev.filter((_, i) => i !== index));
+    setImagePreviewUrls(prev => {
+      URL.revokeObjectURL(prev[index]); // 메모리 정리
+      return prev.filter((_, i) => i !== index);
+    });
+  };
+
+  const resetImages = () => {
+    // 메모리 정리
+    imagePreviewUrls.forEach(url => URL.revokeObjectURL(url));
+    setSelectedImages([]);
+    setImagePreviewUrls([]);
+  };
+
+  return {
+    selectedImages,
+    imagePreviewUrls,
+    handleImageChange,
+    removeImage,
+    resetImages
+  };
+}
